requirements.txt
annotated-types==0.7.0
anyio==3.7.1
Authlib==1.5.2
blis==0.7.11
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.3
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.105.0
filelock==3.18.0
fsspec==2025.3.2
gensim==4.3.3
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
huggingface-hub==0.15.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
ko-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_lg-3.7.0/ko_core_news_lg-3.7.0-py3-none-any.whl
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.12
neo4j==5.11.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openpyxl==3.0.9
packaging==25.0
pandas==2.1.3
pillow==11.2.1
preshed==3.0.9
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMuPDF==1.25.5
pypdf==5.5.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.5.1
starlette==0.27.0
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.29.2
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
validators==0.35.0
wasabi==1.1.3
weasel==0.3.4
weaviate-client==3.25.2